backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pandas>=2.2.0
openpyxl>=3.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-deepseek>=0.1.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
cryptography>=42.0.0

